+++ pkg/sdk/client/mock/mock.go
@@ -115,6 +115,36 @@ func (mr *MockOrchestrateClientMockRecorder) GetTxRequest(ctx, txRequestUUID int
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetTxRequest", reflect.TypeOf((*MockOrchestrateClient)(nil).GetTxRequest), ctx, txRequestUUID)
 }
 
+// CallOffTransaction mocks base method
+func (m *MockOrchestrateClient) CallOffTransaction(ctx context.Context, txRequestUUID string) (*api.TransactionResponse, error) {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "CallOffTransaction", ctx, txRequestUUID)
+	ret0, _ := ret[0].(*api.TransactionResponse)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// CallOffTransaction indicates an expected call of CallOffTransaction
+func (mr *MockOrchestrateClientMockRecorder) CallOffTransaction(ctx, txRequestUUID interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CallOffTransaction", reflect.TypeOf((*MockOrchestrateClient)(nil).CallOffTransaction), ctx, txRequestUUID)
+}
+
+// SpeedUpTransaction mocks base method
+func (m *MockOrchestrateClient) SpeedUpTransaction(ctx context.Context, txRequestUUID string, increment *float64) (*api.TransactionResponse, error) {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "SpeedUpTransaction", ctx, txRequestUUID, increment)
+	ret0, _ := ret[0].(*api.TransactionResponse)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// SpeedUpTransaction indicates an expected call of SpeedUpTransaction
+func (mr *MockOrchestrateClientMockRecorder) SpeedUpTransaction(ctx, txRequestUUID, increment interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SpeedUpTransaction", reflect.TypeOf((*MockOrchestrateClient)(nil).SpeedUpTransaction), ctx, txRequestUUID, increment)
+}
+
 // GetSchedule mocks base method
 func (m *MockOrchestrateClient) GetSchedule(ctx context.Context, scheduleUUID string) (*api.ScheduleResponse, error) {
 	m.ctrl.T.Helper()
@@ -787,6 +817,36 @@ func (m *MockTransactionClient) GetTxRequest(ctx context.Context, txRequestUUID 
 func (mr *MockTransactionClientMockRecorder) GetTxRequest(ctx, txRequestUUID interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetTxRequest", reflect.TypeOf((*MockTransactionClient)(nil).GetTxRequest), ctx, txRequestUUID)
+}
+
+// CallOffTransaction mocks base method
+func (m *MockTransactionClient) CallOffTransaction(ctx context.Context, txRequestUUID string) (*api.TransactionResponse, error) {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "CallOffTransaction", ctx, txRequestUUID)
+	ret0, _ := ret[0].(*api.TransactionResponse)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// CallOffTransaction indicates an expected call of CallOffTransaction
+func (mr *MockTransactionClientMockRecorder) CallOffTransaction(ctx, txRequestUUID interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CallOffTransaction", reflect.TypeOf((*MockTransactionClient)(nil).CallOffTransaction), ctx, txRequestUUID)
+}
+
+// SpeedUpTransaction mocks base method
+func (m *MockTransactionClient) SpeedUpTransaction(ctx context.Context, txRequestUUID string, increment *float64) (*api.TransactionResponse, error) {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "SpeedUpTransaction", ctx, txRequestUUID, increment)
+	ret0, _ := ret[0].(*api.TransactionResponse)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// SpeedUpTransaction indicates an expected call of SpeedUpTransaction
+func (mr *MockTransactionClientMockRecorder) SpeedUpTransaction(ctx, txRequestUUID, increment interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SpeedUpTransaction", reflect.TypeOf((*MockTransactionClient)(nil).SpeedUpTransaction), ctx, txRequestUUID, increment)
 }
 
 // MockScheduleClient is a mock of ScheduleClient interface