+++ services/api/service/controllers/transactions.go
@@ -3,8 +3,10 @@ package controllers
 import (
 	"encoding/json"
 	"net/http"
+	"strconv"
 
 	"github.com/consensys/orchestrate/pkg/types/api"
+	"github.com/consensys/quorum-key-manager/pkg/errors"
 
 	jsonutils "github.com/consensys/orchestrate/pkg/encoding/json"
 	"github.com/consensys/orchestrate/pkg/toolkit/app/http/httputil"
@@ -19,6 +21,7 @@ var _ entities.ETHTransactionParams
 
 const (
 	IdempotencyKeyHeader = "X-Idempotency-Key"
+	MinBoostValue        = 0.1
 )
 
 type TransactionsController struct {
@@ -43,6 +46,10 @@ func (c *TransactionsController) Append(router *mux.Router) {
 		Handler(http.HandlerFunc(c.deployContract))
 	router.Methods(http.MethodGet).Path("/transactions/{uuid}").
 		Handler(http.HandlerFunc(c.getOne))
+	router.Methods(http.MethodPut).Path("/transactions/{uuid}/speed-up").
+		Handler(http.HandlerFunc(c.speedUp))
+	router.Methods(http.MethodPut).Path("/transactions/{uuid}/call-off").
+		Handler(http.HandlerFunc(c.callOff))
 	router.Methods(http.MethodGet).Path("/transactions").
 		Handler(http.HandlerFunc(c.search))
 }
@@ -268,4 +275,70 @@ func (c *TransactionsController) search(rw http.ResponseWriter, request *http.Re
 	}
 
 	_ = json.NewEncoder(rw).Encode(response)
+}
+
+// @Summary      Speed up transaction timeIncrease transaction gas price
+// @Description  Speed up transaction time by an increase its gas price
+// @Tags         Transactions
+// @Produce      json
+// @Security     ApiKeyAuth
+// @Security     JWTAuth
+// @Param        uuid  path      string               true  "UUID of the transaction request"
+// @Param        boost  query     float64              false  "gas price increment percentage, default value is 0.05 (5%)"
+// @Failure      404   {object}  infra.ErrorResponse  "Transaction request not found"
+// @Failure      500   {object}  infra.ErrorResponse  "Internal server error"
+// @Router       /transactions/{uuid}/speed-up [put]
+func (c *TransactionsController) speedUp(rw http.ResponseWriter, req *http.Request) {
+	rw.Header().Set("Content-Type", "application/json")
+	ctx := req.Context()
+	var boostValue = MinBoostValue // Default
+	var err error
+
+	uuid := mux.Vars(req)["uuid"]
+	boostInput := req.URL.Query().Get("boost")
+	if boostInput != "" {
+		boostValue, err = strconv.ParseFloat(boostInput, 64)
+		if err != nil {
+			httputil.WriteHTTPErrorResponse(rw, errors.InvalidFormatError("expected float as incrementing value"))
+			return
+		}
+		if boostValue < MinBoostValue {
+			err = errors.InvalidFormatError("minimum boost value is %f", MinBoostValue)
+			httputil.WriteHTTPErrorResponse(rw, err)
+			return
+		}
+	}
+
+	tx, err := c.ucs.SpeedUp().Execute(ctx, uuid, boostValue, multitenancy.UserInfoValue(ctx))
+	if err != nil {
+		httputil.WriteHTTPErrorResponse(rw, err)
+		return
+	}
+
+	_ = json.NewEncoder(rw).Encode(formatters.FormatTxResponse(tx))
+}
+
+// @Summary      Call off transaction
+// @Description  Call off transaction by sending empty data transaction with same nonce and additional 10% more gas
+// @Tags         Transactions
+// @Produce      json
+// @Security     ApiKeyAuth
+// @Security     JWTAuth
+// @Param        uuid  path      string               true  "UUID of the transaction request"
+// @Failure      404   {object}  infra.ErrorResponse  "Transaction request not found"
+// @Failure      500   {object}  infra.ErrorResponse  "Internal server error"
+// @Router       /transactions/{uuid}/call-off [put]
+func (c *TransactionsController) callOff(rw http.ResponseWriter, request *http.Request) {
+	rw.Header().Set("Content-Type", "application/json")
+	ctx := request.Context()
+
+	uuid := mux.Vars(request)["uuid"]
+
+	tx, err := c.ucs.CallOff().Execute(ctx, uuid, multitenancy.UserInfoValue(ctx))
+	if err != nil {
+		httputil.WriteHTTPErrorResponse(rw, err)
+		return
+	}
+
+	_ = json.NewEncoder(rw).Encode(formatters.FormatTxResponse(tx))
 }